add jsdoc for clarity
