+++ src/json-processor.js
@@ -1,12 +1,57 @@
+import fs from 'fs';
+/**
+ * @description Helper class that processes uploaded json files
+ * or raw objects
+ */
 export default class JSONProcessor {
-
-  static process(body, iIndex) {
+  /**
+   * 
+   * @param {object} body - javascript json object
+   * @param {InvertedIndex} iIndex - An instance of InvertedIndex class
+   * @return {object|string}  - An object containing generated indices or
+   * error message
+   */
+  static processRaw(body, iIndex) {
     const fileNames = Object.keys(body);
     let index;
     fileNames.forEach((fileName) => {
       const fileContent = body[fileName];
       index = iIndex.createIndex(fileName, fileContent);
     });
+    return index;
+  }
+
+  /**
+   * @description - Helper class that helps inverted-index api to 
+   * file uploads
+   * @param {Array} files - file fields generated by multer
+   * @param {InvertedIndex} iIndex -An instance of inverted index
+   * @returns {object|string} - Generated index or erro message
+   */
+  static processFiles(files, iIndex) {
+    let books;
+    let index;
+    try {
+      files.forEach((file) => {
+        const fileType = file.mimetype.toLowerCase();
+        if (fileType === 'application/json') {
+          const filePath = `./dist/uploads/${file.filename}`;
+          const options = { encoding: 'utf-8' };
+          books = fs.readFileSync(filePath, options);
+          books = books.toString();
+          try {
+            books = JSON.parse(books);
+            index = iIndex.createIndex(file.filename, books);
+          } catch (error) {
+            throw new Error('malformed json file');
+          }
+        } else {
+          throw new Error('invalid file: upload json file');
+        }
+      });
+    } catch (error) {
+      index = error.message;
+    }
     return index;
   }
 }
